Created the Init Structure
Created the project as no standalone, with default values when creating
the project.

Added Bootstrap and Bootstrap icons to the project, created the admin
area.

And configured the routing from app routing to it, and made sure that it
is working.

Implemented the login form for the admin and made it so that it must be
authenticated to go to the admin pages, and if he is already logged in,
then he will be redirected to the dashboard page.

Added the most of the necessary services and modules and created the
admin layout.

Now working on implementing the real functionality.
Created the project as no standalone, with default values when creating the project.

Added Bootstrap and Bootstrap icons to the project, created the admin area.
And configured the routing from app routing to it, and made sure that it is working.

Implemented the login form for the admin and made it so that it must be authenticated to go to the admin pages, and if he is already logged in, then he will be redirected to the dashboard page.

Added the most of the necessary services and modules and created the admin layout.

Now working on implementing the real functionality.